requirements.txt
playwright==1.44.0
motor==3.5.0
pymongo==4.7.2
python-dotenv==1.0.1
httpx==0.27.2
